requirements.txt
flask==3.0.2
flask-cors==4.0.0
werkzeug==3.1.3
jinja2==3.1.3
itsdangerous==2.2.0
click==8.1.7
blinker==1.9.0
python-dotenv==1.0.1
beautifulsoup4==4.12.2
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.3.0
spacy==3.6.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl
networkx==3.1
matplotlib==3.7.2
nltk==3.8.1
transformers==4.35.0
pdfkit==1.0.0
psutil==5.9.5
juriscraper>=2.6.0
PyPDF2==3.0.1
setuptools>=68.0.0
wheel>=0.40.0
redis>=4.0.0
aiohttp==3.9.3
pyjwt==2.8.0
uvicorn==0.27.1
fastapi==0.110.0
python-multipart==0.0.9
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.6.3
pydantic-settings==2.2.1
httpx==0.27.0
sqlalchemy==2.0.28
alembic==1.13.1
bcrypt==4.1.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
email-validator==2.1.1